+++ plone/app/event/browser/event_view.py
@@ -91,14 +91,45 @@ class EventView(BrowserView):
 
     @property
     def occurrences(self):
-        events = []
+        """Returns all occurrences for this context. The maximum
+        defaults to 7 occurrences. If there are more occurrences
+        defined for this context, the result will contain the last item
+        of the occurrence list.
+
+        :rtype: dict - with ``events`` and ``tail`` as keys.
+        """
+        eventsinfo = dict(events=[], tail=None)
         context = self.context
-        occurrences = IRecurrenceSupport(context, None)
-        if occurrences is not None:
-            for occ in occurrences.occurrences():
-                acc = IEventAccessor(occ)
-                display = prepare_for_display(
-                    self.context, acc.start, acc.end, acc.whole_day)
-                display.update({'url': acc.context.absolute_url()})
-                events.append(display)
-        return events
+        adapter = IRecurrenceSupport(context, None)
+        if adapter is not None:
+            occurrences = adapter.occurrences()
+            eventsinfo['events'], eventsinfo['tail'] = (
+                self._get_occurrences_helper(occurrences)
+            )
+        return eventsinfo
+
+    def _get_occurrences_helper(self, occ_list, limit=7):
+        """For many occurrences we limit the amount of occurrences to
+        display. That is, this method returns the first 6 (limit)
+        occurrences and the last occurrence in the list.
+
+        :param occ_list: The list of occurrences returned from
+                         IRecurrenceSupport
+        :type occ_list: list
+        :param limit: optional, defaults to 7
+        :type limit: integer
+        :rtype: tuple of (list of events, last item of occ_list)
+        """
+        events = []
+        occurrences = occ_list[:limit]
+        occurrences.append(occ_list.pop())
+        many = len(occ_list) > limit
+        for occ in occurrences:
+            acc = IEventAccessor(occ)
+            display = prepare_for_display(
+                self.context, acc.start, acc.end, acc.whole_day)
+            display.update({'url': acc.context.absolute_url()})
+            events.append(display)
+        tail = events.pop()
+        tail = tail if many else None
+        return (events, tail)