+++ plone/app/event/tests/test_base_module.py
@@ -1,24 +1,26 @@
 import datetime
 import pytz
 import unittest2 as unittest
+import zope.component
 from DateTime import DateTime
-from plone.app.event.base import DEFAULT_END_DELTA
-from plone.app.event.base import DT
-from plone.app.event.base import default_end_DT
-from plone.app.event.base import default_end_dt
-from plone.app.event.base import default_start_DT
-from plone.app.event.base import default_start_dt
-from plone.app.event.base import localized_now
-from plone.app.event.base import get_portal_events
-from plone.app.event.base import default_timezone
+from plone.app.event.base import (
+    DEFAULT_END_DELTA,
+    DT,
+    default_end_DT,
+    default_end_dt,
+    default_start_DT,
+    default_start_dt,
+    default_timezone,
+    get_occurrences,
+    get_portal_events,
+    localized_now
+)
+from plone.app.event.interfaces import IEventSettings
+from plone.app.event.testing import PAEventAT_INTEGRATION_TESTING
+from plone.app.event.testing import PAEvent_INTEGRATION_TESTING
 from plone.app.testing import TEST_USER_ID
 from plone.app.testing import setRoles
-import zope.component
 from plone.registry.interfaces import IRegistry
-from plone.app.event.interfaces import IEventSettings
-
-from plone.app.event.testing import PAEvent_INTEGRATION_TESTING
-from plone.app.event.testing import PAEventAT_INTEGRATION_TESTING
 
 
 class TestBaseModule(unittest.TestCase):
@@ -64,27 +66,39 @@ class TestBaseModule(unittest.TestCase):
 
     def test_DT(self):
         # Python datetime with valid zone. Zope converts it to GMT+1...
+        # TODO: DateTime better shouldn't do this!
         cet = pytz.timezone('CET')
-        self.assertTrue(DT(datetime.datetime(2011, 11, 11, 11, 0, 0, tzinfo=cet)) ==
-                        DateTime('2011/11/11 11:00:00 GMT+1'))
+        self.assertTrue(
+            DT(datetime.datetime(2011, 11, 11, 11, 0, 0, tzinfo=cet)) ==
+            DateTime('2011/11/11 11:00:00 GMT+1')
+        )
 
         # Python dates get converted to a DateTime with timecomponent including
         # a timezone
-        self.assertTrue(DT(datetime.date(2011, 11, 11)) ==
-                        DateTime('2011/11/11 00:00:00 UTC'))
+        self.assertTrue(
+            DT(datetime.date(2011, 11, 11)) ==
+            DateTime('2011/11/11 00:00:00 UTC')
+        )
 
         # DateTime with valid zone
-        self.assertTrue(DT(DateTime(2011, 11, 11, 11, 0, 0, 'Europe/Vienna'))
-                        == DateTime('2011/11/11 11:00:00 Europe/Vienna'))
+        self.assertTrue(
+            DT(DateTime(2011, 11, 11, 11, 0, 0, 'Europe/Vienna')) ==
+            DateTime('2011/11/11 11:00:00 Europe/Vienna')
+        )
 
         # Zope DateTime with valid DateTime zone but invalid pytz is kept as is
-        self.assertTrue(DT(DateTime(2011, 11, 11, 11, 0, 0, 'GMT+1')) ==
-                        DateTime('2011/11/11 11:00:00 GMT+1'))
+        self.assertTrue(
+            DT(DateTime(2011, 11, 11, 11, 0, 0, 'GMT+1')) ==
+            DateTime('2011/11/11 11:00:00 GMT+1')
+        )
 
         # Invalid datetime zones are converted to the portal timezone
         # Testing with no timezone
-        self.assertTrue(DT(datetime.datetime(2011, 11, 11, 11, 0, 0)) ==
-                        DateTime('2011/11/11 11:00:00 UTC'))
+        self.assertTrue(
+            DT(datetime.datetime(2011, 11, 11, 11, 0, 0)) ==
+            DateTime('2011/11/11 11:00:00 UTC')
+        )
+
 
 
 class TestBaseModuleQueryPydt(unittest.TestCase):
@@ -206,6 +220,8 @@ class TestBaseModuleQueryPydt(unittest.TestCase):
                                  range_end=self.now)
         self.assertTrue(len(res9) == 3)
 
+    def test_get_occurrences(self):
+        get_occurrences(object, [], range_start=datetime.datetime.today())
 
 class TestBaseModuleQueryZDT(unittest.TestCase):
     layer = PAEventAT_INTEGRATION_TESTING